3.1-drf-intro/smart_home/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
